restructuring part a
