+++ modules/Layers.py
@@ -1,4 +1,8 @@
-
+'''
+Needs some cleaning.
+On the longer term, let's keep this just a wrapper module for layers,
+but the layers themselves to other files
+'''
 
 # Define custom layers here and add them to the global_layers_list dict (important!)
 global_layers_list = {}
@@ -29,9 +33,6 @@ global_layers_list['CondensateToPseudoRS']=CondensateToPseudoRS
 
 from LayersRagged import GridMaxPoolReduction
 global_layers_list['GridMaxPoolReduction']=GridMaxPoolReduction
-
-from LayersRagged import RaggedConstructTensor
-global_layers_list['RaggedConstructTensor']=RaggedConstructTensor
 
 from LayersRagged import RaggedGlobalExchange
 global_layers_list['RaggedGlobalExchange']=RaggedGlobalExchange
@@ -169,14 +170,14 @@ from GravNetLayersRagged import EdgeConvStatic
 global_layers_list['EdgeConvStatic']=EdgeConvStatic
 
 ### odd debug layers
-from debugLayers import PlotCoordinates
+from DebugLayers import PlotCoordinates
 global_layers_list['PlotCoordinates']=PlotCoordinates
 
 
-from lossLayers import LLNotNoiseClassifier,CreateTruthSpectatorWeights
-from lossLayers import LLLocalClusterCoordinates, LLClusterCoordinates,LLFillSpace
-from lossLayers import LossLayerBase, LLFullObjectCondensation,LLNeighbourhoodClassifier
-from lossLayers import LLEdgeClassifier
+from LossLayers import LLNotNoiseClassifier,CreateTruthSpectatorWeights
+from LossLayers import LLLocalClusterCoordinates, LLClusterCoordinates,LLFillSpace
+from LossLayers import LossLayerBase, LLFullObjectCondensation,LLNeighbourhoodClassifier
+from LossLayers import LLEdgeClassifier
 import traceback
 import os
 
@@ -207,7 +208,7 @@ global_layers_list['MeanMaxDistanceRegularizer']=MeanMaxDistanceRegularizer
 
 
 #also this one needs to be passed
-from initializers import EyeInitializer
+from Initializers import EyeInitializer
 global_layers_list['EyeInitializer']=EyeInitializer
 
 ####### some implementations
@@ -217,7 +218,6 @@ global_layers_list['EyeInitializer']=EyeInitializer
 from tensorflow.keras.layers import Layer
 import tensorflow.keras.backend as K
 import tensorflow as tf
-from Loss_tools import deltaPhi
 
 
 class GausActivation(Layer):
@@ -294,37 +294,6 @@ class ExpMinusOne(Layer):
 global_layers_list['ExpMinusOne']=ExpMinusOne
 
 
-class CenterPhi(Layer):
-    '''
-    Centers phi to the first input vertex, such that the 2pi modulo behaviour
-    disappears for a small selection
-    '''
-    def __init__(self, phi_feature_index, **kwargs):
-        super(CenterPhi, self).__init__(**kwargs)
-        self.phi_feature_index=phi_feature_index
-
-    def compute_output_shape(self, input_shape):
-        return input_shape
-
-    def call(self, inputs):
-        phi = inputs[...,self.phi_feature_index:self.phi_feature_index+1]
-
-        reference = inputs[...,0:1,self.phi_feature_index:self.phi_feature_index+1]
-
-        n_phi = deltaPhi( reference, phi )
-
-        rest_left  = inputs[...,:self.phi_feature_index]
-        rest_right = inputs[...,self.phi_feature_index+1:]
-
-        return tf.concat( [rest_left, n_phi,rest_right ] , axis=-1 )
-
-    def get_config(self):
-        config = {'phi_feature_index': self.phi_feature_index}
-        base_config = super(CenterPhi, self).get_config()
-        return dict(list(base_config.items()) + list(config.items() ))
-
-
-global_layers_list['CenterPhi']=CenterPhi
 
 
 class CreateZeroMask(Layer):
