+++ modules/Loss_implementations.py
@@ -1,5 +1,6 @@
 
 print(">>>> WARNING: THE MODULE", __name__ ,"IS MARKED FOR REMOVAL")
+raise ImportError("MODULE",__name__,"will be removed")
 
 '''
 This file should contain the actual loss implementations.
