+++ modules/caloGraphNN_keras.py
@@ -1,5 +1,6 @@
 
 print(">>>> WARNING: THE MODULE", __name__ ,"IS MARKED FOR REMOVAL")
+raise ImportError("MODULE",__name__,"will be removed")
 
 import tensorflow as tf
 import keras
