+++ modules/globals.py
@@ -1,4 +1,6 @@
 
-#reserved for global constants used by multiple modules
+'''
+file reserved for global constants/defaults used by multiple modules
+'''
 
 removed_noise_coordinate = 100.
