+++ modules/hgcal_predictor.py
@@ -1,4 +1,7 @@
 
+'''
+JK: What are all these useless imports?
+'''
 from argparse import ArgumentParser
 from Layers import RobustModel
 import tensorflow.python.keras.saving.saved_model.load
