+++ modules/training_metrics_plots.py
@@ -6,6 +6,7 @@ import plotly.graph_objects as go
 from plotly.subplots import make_subplots
 import math
 
+print("MODULE OBSOLETE?",__name__)
 
 _debug = True
 class TrainingMetricPlots():
